Refactoring to OAuth2.0 server module
