+++ .idea/encodings.xml
@@ -1,6 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project version="4">
   <component name="Encoding">
+    <file url="file://$PROJECT_DIR$/oauth-authorization-server/src/main/java" charset="UTF-8" />
     <file url="file://$PROJECT_DIR$/outh2.0-auth-server/src/main/java" charset="UTF-8" />
     <file url="file://$PROJECT_DIR$/spring-security-client/src/main/java" charset="UTF-8" />
   </component>
