+++ oauth-authorization-server/pom.xml
@@ -9,14 +9,19 @@
 		<relativePath/> <!-- lookup parent from repository -->
 	</parent>
 	<groupId>np.com.rabin</groupId>
-	<artifactId>outh2.0-auth-server</artifactId>
+	<artifactId>oauth-authorization-server</artifactId>
 	<version>0.0.1-SNAPSHOT</version>
-	<name>outh2.0-auth-server</name>
+	<name>oauth-authorization-server</name>
 	<description>OAuth Spring Boot</description>
 	<properties>
 		<java.version>1.8</java.version>
 	</properties>
 	<dependencies>
+		<dependency>
+			<groupId>org.springframework.security</groupId>
+			<artifactId>spring-security-oauth2-authorization-server</artifactId>
+			<version>0.2.2</version>
+		</dependency>
 		<dependency>
 			<groupId>org.springframework.boot</groupId>
 			<artifactId>spring-boot-starter-data-jpa</artifactId>
