+++ oauth-authorization-server/src/main/resources/application.properties
@@ -1,0 +1,12 @@
+
+server.port = 8081
+
+spring.datasource.url = jdbc:mysql://localhost:3306/user-reg
+spring.datasource.username = root
+spring.datasource.password = P@ss2021
+spring.jpa.format-sql = true
+spring.jpa.show-sql = true
+#spring.jpa.hibernate.ddl-auto = update
+
+spring.datasource.driver-class-name = com.mysql.cj.jdbc.Driver
+
