+++ oauth-authorization-server/src/test/java/np/com/rabin/oauthauthorizationserver/OauthAuthorizationServerApplicationTests.java
@@ -1,10 +1,10 @@
-package np.com.rabin.oauthserver;
+package np.com.rabin.oauthauthorizationserver;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
 @SpringBootTest
-class ApplicationTests {
+class OauthAuthorizationServerApplicationTests {
 
 	@Test
 	void contextLoads() {
