+++ spring-security-client/pom.xml
@@ -17,6 +17,8 @@
 		<java.version>8</java.version>
 	</properties>
 	<dependencies>
+
+
 		<dependency>
 			<groupId>org.springframework.boot</groupId>
 			<artifactId>spring-boot-starter-data-jpa</artifactId>
